refactor(luascript): move lua env to own file
