+++ src/actions.cpp
@@ -10,6 +10,7 @@
 #include "game.h"
 #include "housetile.h"
 #include "lua/meta.h"
+#include "lua/env.h"
 #include "spells.h"
 
 extern Game g_game;
@@ -383,7 +384,7 @@ bool Action::executeUse(const std::shared_ptr<Player>& player, const std::shared
 		return false;
 	}
 
-	ScriptEnvironment* env = tfs::lua::getScriptEnv();
+	const auto env = tfs::lua::getScriptEnv();
 	env->setScriptId(scriptId, scriptInterface);
 
 	lua_State* L = scriptInterface->getLuaState();
