+++ src/creatureevent.cpp
@@ -6,6 +6,7 @@
 #include "creatureevent.h"
 
 #include "item.h"
+#include "lua/env.h"
 #include "lua/error.h"
 #include "lua/meta.h"
 
@@ -284,7 +285,7 @@ bool CreatureEvent::executeOnLogin(const std::shared_ptr<Player>& player) const
 		return false;
 	}
 
-	ScriptEnvironment* env = tfs::lua::getScriptEnv();
+	const auto env = tfs::lua::getScriptEnv();
 	env->setScriptId(scriptId, scriptInterface);
 
 	lua_State* L = scriptInterface->getLuaState();
@@ -303,7 +304,7 @@ bool CreatureEvent::executeOnLogout(const std::shared_ptr<Player>& player) const
 		return false;
 	}
 
-	ScriptEnvironment* env = tfs::lua::getScriptEnv();
+	const auto env = tfs::lua::getScriptEnv();
 	env->setScriptId(scriptId, scriptInterface);
 
 	lua_State* L = scriptInterface->getLuaState();
@@ -322,7 +323,7 @@ void CreatureEvent::executeOnReconnect(const std::shared_ptr<Player>& player) co
 		return;
 	}
 
-	ScriptEnvironment* env = tfs::lua::getScriptEnv();
+	const auto env = tfs::lua::getScriptEnv();
 	env->setScriptId(scriptId, scriptInterface);
 
 	lua_State* L = scriptInterface->getLuaState();
@@ -341,7 +342,7 @@ bool CreatureEvent::executeOnThink(const std::shared_ptr<Creature>& creature, ui
 		return false;
 	}
 
-	ScriptEnvironment* env = tfs::lua::getScriptEnv();
+	const auto env = tfs::lua::getScriptEnv();
 	env->setScriptId(scriptId, scriptInterface);
 
 	lua_State* L = scriptInterface->getLuaState();
@@ -363,7 +364,7 @@ bool CreatureEvent::executeOnPrepareDeath(const std::shared_ptr<Creature>& creat
 		return false;
 	}
 
-	ScriptEnvironment* env = tfs::lua::getScriptEnv();
+	const auto env = tfs::lua::getScriptEnv();
 	env->setScriptId(scriptId, scriptInterface);
 
 	lua_State* L = scriptInterface->getLuaState();
@@ -394,7 +395,7 @@ bool CreatureEvent::executeOnDeath(const std::shared_ptr<Creature>& creature, co
 		return false;
 	}
 
-	ScriptEnvironment* env = tfs::lua::getScriptEnv();
+	const auto env = tfs::lua::getScriptEnv();
 	env->setScriptId(scriptId, scriptInterface);
 
 	lua_State* L = scriptInterface->getLuaState();
@@ -434,7 +435,7 @@ bool CreatureEvent::executeAdvance(const std::shared_ptr<Player>& player, skills
 		return false;
 	}
 
-	ScriptEnvironment* env = tfs::lua::getScriptEnv();
+	const auto env = tfs::lua::getScriptEnv();
 	env->setScriptId(scriptId, scriptInterface);
 
 	lua_State* L = scriptInterface->getLuaState();
@@ -457,7 +458,7 @@ void CreatureEvent::executeOnKill(const std::shared_ptr<Creature>& creature, con
 		return;
 	}
 
-	ScriptEnvironment* env = tfs::lua::getScriptEnv();
+	const auto env = tfs::lua::getScriptEnv();
 	env->setScriptId(scriptId, scriptInterface);
 
 	lua_State* L = scriptInterface->getLuaState();
@@ -479,7 +480,7 @@ void CreatureEvent::executeModalWindow(const std::shared_ptr<Player>& player, ui
 		return;
 	}
 
-	ScriptEnvironment* env = tfs::lua::getScriptEnv();
+	const auto env = tfs::lua::getScriptEnv();
 	env->setScriptId(scriptId, scriptInterface);
 
 	lua_State* L = scriptInterface->getLuaState();
@@ -504,7 +505,7 @@ bool CreatureEvent::executeTextEdit(const std::shared_ptr<Player>& player, const
 		return false;
 	}
 
-	ScriptEnvironment* env = tfs::lua::getScriptEnv();
+	const auto env = tfs::lua::getScriptEnv();
 	env->setScriptId(scriptId, scriptInterface);
 
 	lua_State* L = scriptInterface->getLuaState();
@@ -543,7 +544,7 @@ void CreatureEvent::executeHealthChange(const std::shared_ptr<Creature>& creatur
 		return;
 	}
 
-	ScriptEnvironment* env = tfs::lua::getScriptEnv();
+	const auto env = tfs::lua::getScriptEnv();
 	env->setScriptId(scriptId, scriptInterface);
 
 	lua_State* L = scriptInterface->getLuaState();
@@ -587,7 +588,7 @@ void CreatureEvent::executeManaChange(const std::shared_ptr<Creature>& creature,
 		return;
 	}
 
-	ScriptEnvironment* env = tfs::lua::getScriptEnv();
+	const auto env = tfs::lua::getScriptEnv();
 	env->setScriptId(scriptId, scriptInterface);
 
 	lua_State* L = scriptInterface->getLuaState();
@@ -626,7 +627,7 @@ void CreatureEvent::executeExtendedOpcode(const std::shared_ptr<Player>& player,
 		return;
 	}
 
-	ScriptEnvironment* env = tfs::lua::getScriptEnv();
+	const auto env = tfs::lua::getScriptEnv();
 	env->setScriptId(scriptId, scriptInterface);
 
 	lua_State* L = scriptInterface->getLuaState();
