+++ src/monster.cpp
@@ -113,7 +113,7 @@ void Monster::onCreatureAppear(const std::shared_ptr<Creature>& creature, bool, 
 			return;
 		}
 
-		ScriptEnvironment* env = tfs::lua::getScriptEnv();
+		const auto env = tfs::lua::getScriptEnv();
 		env->setScriptId(mType->info.creatureAppearEvent, scriptInterface);
 
 		lua_State* L = scriptInterface->getLuaState();
@@ -155,7 +155,7 @@ void Monster::onRemoveCreature(const std::shared_ptr<Creature>& creature, bool i
 			return;
 		}
 
-		ScriptEnvironment* env = tfs::lua::getScriptEnv();
+		const auto env = tfs::lua::getScriptEnv();
 		env->setScriptId(mType->info.creatureDisappearEvent, scriptInterface);
 
 		lua_State* L = scriptInterface->getLuaState();
@@ -197,7 +197,7 @@ void Monster::onCreatureMove(const std::shared_ptr<Creature>& creature, const st
 			return;
 		}
 
-		ScriptEnvironment* env = tfs::lua::getScriptEnv();
+		const auto env = tfs::lua::getScriptEnv();
 		env->setScriptId(mType->info.creatureMoveEvent, scriptInterface);
 
 		lua_State* L = scriptInterface->getLuaState();
@@ -279,7 +279,7 @@ void Monster::onCreatureSay(const std::shared_ptr<Creature>& creature, SpeakClas
 			return;
 		}
 
-		ScriptEnvironment* env = tfs::lua::getScriptEnv();
+		const auto env = tfs::lua::getScriptEnv();
 		env->setScriptId(mType->info.creatureSayEvent, scriptInterface);
 
 		lua_State* L = scriptInterface->getLuaState();
@@ -698,7 +698,7 @@ void Monster::onThink(uint32_t interval)
 		}
 
 		LuaScriptInterface* scriptInterface = mType->info.scriptInterface;
-		ScriptEnvironment* env = tfs::lua::getScriptEnv();
+		const auto env = tfs::lua::getScriptEnv();
 		env->setScriptId(mType->info.thinkEvent, scriptInterface);
 
 		lua_State* L = scriptInterface->getLuaState();
