+++ src/talkaction.cpp
@@ -126,7 +126,7 @@ bool TalkAction::executeSay(const std::shared_ptr<Player>& player, const std::st
 		return false;
 	}
 
-	ScriptEnvironment* env = tfs::lua::getScriptEnv();
+	const auto env = tfs::lua::getScriptEnv();
 	env->setScriptId(scriptId, scriptInterface);
 
 	lua_State* L = scriptInterface->getLuaState();
